requirements.txt
fastapi
uvicorn
pydantic
numpy
tensorflow
transformers
gdown
python-dotenv
sqlalchemy
psycopg2-binary
python-multipart
alembic
pandas